requirements.txt
dgl==0.7.2
tensorflow-addons==0.14.0
tensorflow-gpu==2.4.1
pandas==1.1.5
scikit-learn==0.24.2
tqdm==4.62.3
scipy==1.5.3